requirements.txt
numpy>=1.26,<3
pandas>=2.2,<3
matplotlib>=3.8,<4
opencv-python>=4.8,<5
h5py>=3.9,<4
PyYAML>=6,<7
tqdm>=4.66,<5
snntorch>=0.8,<0.10
# Instala torch/torchvision según CUDA (p.ej. cu129 para tu 4080)
